fix(client): whitelist class prefix selectors in PurgeCSS
PurgeCSS is unable to handle the class-prefix selectors that Cirrus uses
(of the form `[class*=" btn-"], [class^="btn-"]`), so whitelist them
manually to ensure they are not pruned.

PurgeCSS has basic support for attribute selectors which test for an
attribute's existence, but does not recognize any value comparisons in
the selectors, so we must whitelist all `class` selectors as we have no
ability to whitelist based on what the selector is actually selecting
for.