feat(server): parse real IP from Cloudflare headers
Since Fastify does not have any hooks that run before the request is
logged, we have to override the request serializer to modify `req.ip` in
the logged request data. We also register a hook as early as possible to
modify the real `req.ip`. The serializer is copied from the default
implemenation in Fastify.