feat(server): implement route handler helpers
Implement `makeFastifyRoute` to create a fastify route object from a
route object and a handler function.

The handler function takes the same arguments as before, except with an
added `res` property which contains factory functions for all valid
response kinds.

Schema transformer for fast-json-schema removes the entire schema when
there is a `oneOf` at the top level with `rawJson`, because
fast-json-schema does not support it, and pushes the `oneOf` down for
standard format responses.