fix(server): ensure failure responses are in schemas
For response kinds which are not part of the declared route (`badBody`
and `errorInternal`, which both should not occur when the API is used
correctly and is functioning correctly, and thus aren't included in the
definitions for generating types), ensure they are added to the response
schema fed to fast-json-stringify when appropriate.

Without this, such implicit kinds which contain `data` can fail to
serialize correctly.