refactor(server): derive handler type off of FastifyInstance
When casting the handler functions exported from individual routes,
derive the target type by inspecting the `route` method of the
FastifyInstance we are registering them on. Since we adjust the instance
type to better match `http.Server`, this fixes a type mismatch that
appears when compiling on some platforms when trying to register the
handlers.