feat(server): rank users with only tiebreakEligible: false
Break ties between users with solves on challenges not marked
`tiebreakEligible: false` the same, but also rank users with only solves
on challenges marked `tiebreakEligible: false` via solve times on those
challenges. This does not change existing rankings, but adds users with
only solves on these challenges to the leaderboard as well.

Co-authored-by: Philip Papurt <ginkoid@gmail.com>