refactor(server): hide leaderboard intermediate info
Remove intermediate calculation bookkeeping data from the types returned
out of the leaderboard worker. This change also strips the data from the
objects, resulting in the creation of a new JS object per leaderboard
entry, but keeps snapshots cleaner. We may wish to consider doing this
in only snapshot serialization in the future.