feat: remove register page
This commit removes the entire register page and user token input
box on the login page. This is a cosmetic change only and does not
change any of the underlying APIs.

With this change, the only way to log in is using Ion or a token
link. This should simplify things and reduce confusion.